+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to this project will be documented in this file.
 - Update babel config and dependencies
 - Upgrade libsodium-wrappers to libsodium-wrappers-sumo
 - Upgrade obs-store to zustand
+- Upgrade bs58check to @scure/base
 - Upgrade node-forge to v1.3.3
 - Upgrade gulp, gulp-zip to jszip
 - Upgrade react-idle-timer, react-circular-progressbar, axios, copy-webpack-plugin, mocha, sass, sass-loader, webpack-cli, @reduxjs/toolkit, sinon, i18next, react, react-dom, react-i18next, react-redux, react-router-dom